+++ inferno-core/CHANGELOG.md
@@ -1,4 +1,10 @@
 # Revision History for inferno-core
 
+## 0.1.0.2 -- 2022-12-31
+* Add `inferno` binary for type-checking and executing inferno programs
+
+## 0.1.0.1 -- 2022-12-1
+* [fix] Function 'year' and 'month' should truncate time properly.
+
 ## 0.1.0.0 -- 2022-11-28
 * Prepare for OSS release